+++ src/cmd_ovlpr/mod.rs
@@ -3,3 +3,4 @@
 pub mod covered;
 pub mod paf2ovlp;
 pub mod replace;
+pub mod restrict;