+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 * Add `ovlpr replace`
+* Add `ovlpr restrict`
 
 ## [0.4.1] - 2019-09-10